@@ -2,8 +2,9 @@
     "name": "printnode/printnode-php",
     "description": "Connect any printer to your application with PrintNode Client and easy to use JSON API",
     "require": {
-        "php": ">=5.3.2",
-        "ext-curl": "*"
+        "php": ">=7.0.0",
+        "ext-curl": "*",
+        "ext-json": "*"
     },
     "keywords": [
         "php",
