@@ -2,16 +2,34 @@
 
 namespace PrintNode;
 
+use BadMethodCallException;
+use DateTime;
+use InvalidArgumentException;
+use function json_last_error;
+use function json_last_error_msg;
+use RuntimeException;
+
+use function get_object_vars;
+use function get_class;
+use function is_array;
+use function is_object;
+use function json_encode;
+use function preg_match;
+use function property_exists;
+use function strtolower;
+use function substr;
+
 /**
  * Entity
- *
  * Base class for entity objects.
  */
 abstract class Entity implements EntityInterface
 {
     /**
      * Recursively cast an object into an array.
+     *
      * @param mixed $object
+     *
      * @return mixed[]
      */
     private static function toArrayRecursive($object)
@@ -19,7 +37,7 @@ abstract class Entity implements EntityInterface
         $output = get_object_vars($object);
 
         foreach ($output as $key => $value) {
-            if ($value instanceof \DateTime) {
+            if ($value instanceof DateTime) {
                 $output[$key] = $value->format('c');
             } elseif (is_object($value)) {
                 $output[$key] = static::toArrayRecursive($value);
@@ -31,25 +49,26 @@ abstract class Entity implements EntityInterface
 
     /**
      * Map array of data to an entity
+     *
      * @param mixed $entityName
      * @param mixed $data
+     *
      * @return Entity
+     * @throws \Exception
      */
     private static function mapDataToEntity($entityName, \stdClass $data)
     {
         $entity = new $entityName();
 
         if (!($entity instanceof Entity)) {
-            throw new \RuntimeException(
+            throw new RuntimeException(
                 sprintf(
                     'Object "%s" must extend Entity',
                     $entityName
                 )
             );
         }
-
         $foreignKeyEntityMap = $entity->foreignKeyEntityMap();
-
         $properties = array_keys(get_object_vars($data));
 
         foreach ($properties as $propertyName) {
@@ -68,8 +87,9 @@ abstract class Entity implements EntityInterface
                     $foreignKeyEntityMap[$propertyName],
                     $data->$propertyName
                 );
-            } elseif (is_string($data->$propertyName) && preg_match('/^\d{4}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2}/', $data->$propertyName)) {
-                $entity->$propertyName = new \DateTime($data->$propertyName);
+            } elseif (is_string($data->$propertyName) &&
+                preg_match('/^\d{4}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2}/', $data->$propertyName)) {
+                $entity->$propertyName = new DateTime($data->$propertyName);
             } else {
                 $entity->$propertyName = json_decode(json_encode($data->$propertyName), true);
             }
@@ -80,7 +100,9 @@ abstract class Entity implements EntityInterface
 
     /**
      * Cast entity into an array
+     *
      * @param void
+     *
      * @return mixed[]
      */
     public function toArray()
@@ -90,24 +112,39 @@ abstract class Entity implements EntityInterface
 
     /**
      * Cast entity into a JSON encoded string
+     *
      * @param void
+     *
      * @return string
      */
     public function __toString()
     {
-        return json_encode($this->toArray());
+        return (string)$this->dataToJson($this->toArray());
+    }
+
+    public function dataToJson($data)
+    {
+        $result = json_encode($data);
+
+        if (json_last_error()) {
+            throw new \RuntimeException(json_last_error_msg(), json_last_error());
+        }
+
+        return $result;
     }
 
     /**
      * Set property on entity
+     *
      * @param mixed $propertyName
      * @param mixed $value
+     *
      * @return void
      */
     public function __set($propertyName, $value)
     {
         if (!property_exists($this, $propertyName)) {
-            throw new \InvalidArgumentException(
+            throw new InvalidArgumentException(
                 sprintf(
                     '%s does not have a property named %s',
                     get_class($this),
@@ -115,19 +152,20 @@ abstract class Entity implements EntityInterface
                 )
             );
         }
-
         $this->$propertyName = $value;
     }
 
     /**
      * Get property on entity
+     *
      * @param mixed $propertyName
+     *
      * @return mixed
      */
     public function __get($propertyName)
     {
         if (!property_exists($this, $propertyName)) {
-            throw new \InvalidArgumentException(
+            throw new InvalidArgumentException(
                 sprintf(
                     '%s does not have a property named %s',
                     get_class($this),
@@ -142,14 +180,16 @@ abstract class Entity implements EntityInterface
     /**
      * Property get/set wrapper for those that prefer
      * $entity->get('propertyName') style access
+     *
      * @param mixed $name
      * @param mixed $arguments
+     *
      * @return mixed
      */
     public function __call($name, $arguments)
     {
         if (!preg_match('/^(get|set)(.+)$/', $name, $matchesArray)) {
-            throw new \BadMethodCallException(
+            throw new BadMethodCallException(
                 sprintf(
                     'method "%s" does not exist on entity "%s"',
                     $name,
@@ -157,13 +197,10 @@ abstract class Entity implements EntityInterface
                 )
             );
         }
-
         $propertyName = $matchesArray[2];
-
-        $propertyName = strtolower(substr($propertyName, 0, 1)). substr($propertyName, 1);
-
+        $propertyName = strtolower(substr($propertyName, 0, 1)) . substr($propertyName, 1);
         if (!property_exists($this, $propertyName)) {
-            throw new \BadMethodCallException(
+            throw new BadMethodCallException(
                 sprintf(
                     'Entity %s does not have a property named %s',
                     get_class($this),
@@ -171,31 +208,27 @@ abstract class Entity implements EntityInterface
                 )
             );
         }
-
         switch ($matchesArray[1]) {
-
             case 'set':
-
                 $this->$propertyName = $arguments[0];
                 break;
-
             case 'get':
-
                 return $this->$propertyName;
-                break;
         }
     }
 
     /**
      * Make an array of specified entity from a Response
-     * @param mixed $entityName
-     * @param Response $response
+     *
+     * @param string       $entityName
+     * @param string|array $content
+     *
      * @return Entity[]
+     * @throws \Exception
      */
     public static function makeFromResponse($entityName, $content)
     {
-        $content = $content;
-        $output = array();
+        $output = [];
         if (is_array($content)) {
             foreach ($content as $entityData) {
                 $output[] = self::makeFromResponse($entityName, $entityData);
