@@ -2,35 +2,54 @@
 
 namespace PrintNode;
 
+use function preg_grep;
+use function preg_match;
+
 /**
  * Response
- *
  * HTTP response object.
  */
 class Response
 {
     /**
      * Original Request URL
+     *
      * @var string
      */
     private $url;
 
     /**
      * Response headers
+     *
      * @var mixed[]
      */
     private $headers;
 
     /**
      * Response body
+     *
      * @var string
      */
     private $content;
 
     /**
+     * @param mixed $url
+     * @param mixed $content
+     * @param mixed $headers
+     */
+    public function __construct($url, $content, array $headers)
+    {
+        $this->url = $url;
+        $this->headers = $headers;
+        $this->content = $content;
+    }
+
+    /**
      * Extract the HTTP status code and message
      * from the Response headers
+     *
      * @param void
+     *
      * @return mixed[]
      */
     private function getStatus()
@@ -43,29 +62,17 @@ class Response
             throw new \RuntimeException('Could not determine HTTP status from API response');
         }
 
-        return array(
+        return [
             'code' => $matchesArray[2],
             'message' => $matchesArray[3],
-        );
-    }
-
-    /**
-     * Constructor
-     * @param mixed $url
-     * @param mixed $content
-     * @param mixed $headers
-     * @return Response
-     */
-    public function __construct($url, $content, array $headers)
-    {
-        $this->url = $url;
-        $this->headers = $headers;
-        $this->content = $content;
+        ];
     }
 
     /**
      * Get Response body
+     *
      * @param void
+     *
      * @return string
      */
     public function getContent()
@@ -75,7 +82,9 @@ class Response
 
     /**
      * Get Response headers
+     *
      * @param void
+     *
      * @return mixed[]
      */
     public function getHeaders()
@@ -85,33 +94,41 @@ class Response
 
     /**
      * Get Response body decoded into an array
+     *
      * @param void
+     *
      * @return mixed
      */
     public function getDecodedContent()
     {
-        return json_decode($this->content, true);
+        return \json_decode($this->content, true);
     }
 
     /**
      * Get HTTP status code
+     *
      * @param void
+     *
      * @return string
      */
     public function getStatusCode()
     {
         $status = $this->getStatus();
+
         return $status['code'];
     }
 
     /**
      * Get HTTP status code
+     *
      * @param void
+     *
      * @return string
      */
     public function getStatusMessage()
     {
         $status = $this->getStatus();
+
         return $status['message'];
     }
 }
